# api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Library API",
    "version": "0.1.0",
    "description": "Catalog and administrator API. Cookie-authenticated writes require an Origin matching LIBRARY_ORIGIN. Use scoped Bearer API keys for automation. Browser sessions are stored in HttpOnly cookies; login responses contain no secret. Nullable PATCH fields distinguish omission from explicit null."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    },
    {
      "apiKey": []
    }
  ],
  "paths": {
    "/api/v1/auth/setup": {
      "get": {
        "operationId": "setupStatus",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupStatus"
                }
              }
            }
          },
          "default": {
            "description": "Operation failed; response includes a safe reason and trace ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": []
      },
      "post": {
        "operationId": "createAdministrator",
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "default": {
            "description": "Operation failed; response includes a safe reason and trace ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "X-Setup-Token",
            "in": "header",
            "required": false,
            "description": "Required when setup status reports requires_token. Read setup-token from the private state directory.",
            "schema": {
              "type": "string",
              "minLength": 43,
              "maxLength": 43
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "trace_id": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": [
              "code",
              "message",
              "trace_id"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "Credentials": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 1024
          }
        },
        "required": [
          "username",
          "password"
        ],
        "additionalProperties": false
      },
      "SetupStatus": {
        "type": "object",
        "properties": {
          "configured": {
            "type": "boolean"
          },
          "requires_token": {
            "type": "boolean",
            "description": "Fresh instances require the private server setup token."
          }
        },
        "required": [
          "configured"
        ]
      },
      "Session": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "scope": {
            "type": "string",
            "enum": [
              "read",
              "manage",
              "admin"
            ]
          },
          "expires_at": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "user_id",
          "scope"
        ]
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "library_session"
      },
      "apiKey": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}